requirements.txt
flask
torch
transformers
# Flask for backend
Flask==2.3.4
flask-cors==3.1.2

# TensorFlow for deepfake model
tensorflow==2.14.0

# PyTorch for Transformers models (CPU version)
torch==2.1.0+cpu
torchvision==0.16.1+cpu
torchaudio==2.1.1+cpu
-f https://download.pytorch.org/whl/cpu/torch_stable.html

# Hugging Face Transformers
transformers==4.41.0
tokenizers==0.14.3

# Newspaper3k for URL extraction
newspaper3k==0.2.8
lxml_html_clean==0.1.6
lxml==4.9.3

# Audio processing
pydub==0.25.1
ffmpeg-python==0.2.2
SpeechRecognition==3.11.0

# Google Gemini AI
google-generativeai==0.1.7

# Environment variables
python-dotenv==1.0.1

# Utilities
numpy==1.25.0
